refactor: remove openraft deps, rename all v2 modules
