+++ storage/src/raft_log_store/block_cache.rs
@@ -2,7 +2,6 @@ use std::sync::Arc;
 
 use futures::Future;
 use moka::future::Cache;
-use tracing::trace;
 
 use super::error::RaftLogStoreError;
 use super::DEFAULT_LOG_BATCH_SIZE;
@@ -28,27 +27,19 @@ impl BlockCache {
         Self { inner: cache }
     }
 
+    #[tracing::instrument(level = "trace", skip(self))]
     pub fn get(&self, file_id: u64, offset: usize) -> Option<Arc<Vec<u8>>> {
-        trace!(
-            file_id = file_id,
-            offset = offset,
-            "try get from block cache:"
-        );
         self.inner.get(&BlockIndex { file_id, offset })
     }
 
+    #[tracing::instrument(level = "trace", skip(self, block))]
     pub async fn insert(&self, file_id: u64, offset: usize, block: Arc<Vec<u8>>) {
-        trace!(
-            file_id = file_id,
-            offset = offset,
-            len = block.len(),
-            "insert to block cache:"
-        );
         self.inner
             .insert(BlockIndex { file_id, offset }, block)
             .await
     }
 
+    #[tracing::instrument(level = "trace", skip(self, f))]
     pub async fn get_or_insert_with<F>(
         &self,
         file_id: u64,
@@ -58,11 +49,6 @@ impl BlockCache {
     where
         F: Future<Output = Result<Arc<Vec<u8>>>>,
     {
-        trace!(
-            file_id = file_id,
-            offset = offset,
-            "get or insert block cache"
-        );
         match self
             .inner
             .get_or_try_insert_with(BlockIndex { file_id, offset }, f)
