--- storage/src/raft_log_store_v2/mod.rs
@@ -1,10 +1,0 @@
-pub mod block_cache;
-pub mod entry;
-pub mod error;
-pub mod log;
-pub mod mem;
-pub mod store;
-
-const DEFAULT_LOG_BATCH_SIZE: usize = 8 << 10;
-
-pub use store::RaftLogStore;
