+++ wheel/src/error.rs
@@ -14,9 +14,7 @@ pub enum Error {
     #[error("serde error: {0}")]
     SerdeError(String),
     #[error("raft error: {0}")]
-    RaftError(#[from] RaftError),
-    #[error("raft error: {0}")]
-    RaftV2Error(#[from] raft::Error),
+    RaftError(#[from] raft::Error),
     #[error("raft manage error: {0}")]
     RaftManagerError(#[from] RaftManageError),
     #[error("meta error: {0}")]
@@ -39,33 +37,9 @@ impl Error {
     pub fn serde_err(e: impl Into<Box<dyn std::error::Error>>) -> Self {
         Self::SerdeError(e.into().to_string())
     }
-
-    pub fn raft_err(e: impl Into<Box<dyn std::error::Error>>) -> Self {
-        RaftError::err(e).into()
-    }
 }
 
 pub type Result<T> = std::result::Result<T, Error>;
-
-#[derive(thiserror::Error, Debug)]
-pub enum RaftError {
-    #[error("raft node not exists: [raft node: {raft_node}] [node: {node}]")]
-    RaftNodeNotExists { raft_node: u64, node: u64 },
-    #[error("raft node already exists: [group: {group}] [raft node: {raft_node}] [node: {node}]")]
-    RaftNodeAlreadyExists {
-        group: u64,
-        raft_node: u64,
-        node: u64,
-    },
-    #[error("other: {0}")]
-    Other(String),
-}
-
-impl RaftError {
-    pub fn err(e: impl Into<Box<dyn std::error::Error>>) -> Self {
-        Self::Other(e.into().to_string())
-    }
-}
 
 #[derive(thiserror::Error, Debug)]
 pub enum RaftManageError {
