+++ wheel/src/worker/mod.rs
@@ -1,4 +1,3 @@
 pub mod heartbeater;
-pub mod kv;
 pub mod raft;
 pub mod sstable_uploader;
